Move AppImage builder functions
